refac : 결제 모달 창 파일 명 변경 및 스타일 개선
