+++ apps/payment/src/components/paymentPage/Address.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { FlexCol, FlexRowsm } from '../styles/layout.style';
 import { BoldBigText } from '../styles/text.style';
-import { InputText, MainButton } from '@repo/ui/components';
-import { useIsMobile } from '../../hooks/useMobile';
+import { BaseButton } from '../styles/product-detail/productInfo.style';
+import { InputText } from '../styles/paymentPage/Address.style';
 
 const Address = ({
   setAddressData,
@@ -12,7 +12,6 @@ const Address = ({
   const [postalCode, setPostalCode] = useState('');
   const [address, setAddress] = useState('');
   const [detailedAddress, setDetailedAddress] = useState('');
-  const isMobile = useIsMobile();
 
   const handleFind = () => {
     new window.daum.Postcode({
@@ -34,24 +33,20 @@ const Address = ({
   return (
     <FlexCol className="items-start justify-start gap-4 mb-10 px-6 sm:px-0">
       <BoldBigText>주소 입력</BoldBigText>
-      <FlexRowsm className="w-full justify-between gap-3">
+      <FlexRowsm className="mt-5 w-full">
         <InputText
-          type={'text'}
+          type="text"
           placeholder="우편번호"
           value={postalCode}
           readOnly
         />
-        <MainButton
-          label={isMobile ? '찾기' : '번호찾기'}
-          className="w-[20%]"
-          textSize={'text-base'}
-          textWeight={'font-bold'}
-          onClick={handleFind}
-        />
+        <BaseButton className="w-[30%] p-2" onClick={handleFind}>
+          번호찾기
+        </BaseButton>
       </FlexRowsm>
       <InputText type="text" placeholder="주소" value={address} readOnly />
       <InputText
-        type={'text'}
+        type="text"
         placeholder="상세주소"
         value={detailedAddress}
         onChange={handleDetailed}
