+++ packages/ui/components/Header.tsx
@@ -151,7 +151,7 @@ export const Header = () => {
             );
           })}
         </div>
-        <FundiButton type="button" onClick={() => navigate('/fundi/request')}>
+        <FundiButton type="button" onClick={() => navigate('/ask-fundi')}>
           <SubTitle>펀디에게 물어보기</SubTitle>
           <img src={fundi} className="w-6" />
         </FundiButton>
